requirements.txt
streamlit>=1.36
geopy>=2.4
requests>=2.32
pandas>=2.0
folium>=0.17
streamlit-folium>=0.20
python-dotenv>=1.0
